final debug dHBV2.0 forward
