+++ deltaModel/__main__.py
@@ -6,12 +6,10 @@ import hydra
 import torch
 from omegaconf import DictConfig
 
-from core.data.data_loaders.loader_hydro_ms import get_dataset_dict
 from core.utils import initialize_config, print_config, set_randomseed
 from core.utils.module_loaders import get_data_loader, get_trainer
+from core.data.data_loaders.loader_hydro_ms import HydroMSDataLoader
 from models.model_handler import ModelHandler as dModel
-
-from core.data.data_loaders.loader_hydro_ms_new import HydroMSDataLoader
 
 log = logging.getLogger(__name__)
 log.setLevel(logging.INFO)
@@ -20,7 +18,7 @@ log.setLevel(logging.INFO)
 @hydra.main(
     version_base='1.3',
     config_path='conf/',
-    config_name='config_ms',
+    config_name='config_ms_dev',
 )
 def main(config: DictConfig) -> None:
     try:
@@ -33,52 +31,30 @@ def main(config: DictConfig) -> None:
         log.info(f"RUNNING MODE: {config['mode']}")
         print_config(config)
 
-
-
-
-
-        # inference_dataset = get_dataset_dict(config, train=False)
-
-        # print("DEBUG-----------------------")
-        # print(inference_dataset['xc_nn_norm'].mean())
-        # print(inference_dataset['x_phy'].mean())
-        # print("DEBUG-----------------------")
-
-
-        data_loader = HydroMSDataLoader(config)
-        inf2 = data_loader.dataset
-        print("DEBUG-----------------------")
-        print(inf2['xc_nn_norm'].mean())
-        print(inf2['x_phy'].mean())
-        print("DEBUG-----------------------")
-
-
-
-
-
-
-        exit()
-
-
         ### Create/Load differentiable model ###
         model = dModel(config, verbose=True)
 
         ### Process datasets ###
         log.info("Processing datasets...")
-        # train_dataset = get_dataset_dict(config, train=True)
-        # eval_dataset = get_dataset_dict(config, train=False)
-        inference_dataset = get_dataset_dict(config, train=False)
+        data_loader = get_data_loader(config['data_loader'])
+        data_loader = data_loader(config, test_split=False, overwrite=False)
 
-        # Trainer
-        from trainers.trainer_ms import Trainer
-        trainer = Trainer(config, model, inf_dataset=inference_dataset, verbose=True)
+        ### Create trainer object ###
+        trainer = get_trainer(config['trainer'])
+        trainer = trainer(
+            config,
+            model,
+            train_dataset = data_loader.train_dataset,
+            eval_dataset = data_loader.eval_dataset,
+            inf_dataset = data_loader.dataset,
+            verbose=True,
+        )
 
         mode = config['mode']
         if mode == 'train':
             trainer.train()
         elif mode == 'test':
-            # trainer.evaluate()
-            trainer.inference()
+            trainer.evaluate()
         elif mode == 'train_test':
             trainer.train()
             trainer.evaluate()
