+++ deltaModel/core/data/__init__.py
@@ -161,198 +161,10 @@ def select_subset(config: Dict,
     return x_tensor.to(config['device']) if torch.cuda.is_available() else x_tensor
 
 
-def get_training_sample(
-    dataset_dict: Dict[str, np.ndarray], 
-    ngrid_train: int,
-    nt: int,
-    config: Dict,
-) -> Dict[str, torch.Tensor]:
-    """Select random sample of data for training batch."""
-    warm_up = config['dpl_model']['phy_model']['warm_up']
-    subset_dims = (config['train']['batch_size'], config['dpl_model']['rho'])
-
-    i_sample, i_t = random_index(ngrid_train, nt, subset_dims, warm_up=warm_up)
-
-    # Remove warmup days for dHBV1.1p...
-    flow_obs = select_subset(config, dataset_dict['target'], i_sample, i_t,
-                             config['dpl_model']['rho'], warm_up=warm_up)
-    
-    # if ('HBV1_1p' in config['dpl_model']['phy_model']['model']) and \
-    # (config['dpl_model']['phy_model']['warm_up_states']) and (config['multimodel_type_type_type'] == 'none'):
-    #     pass
-    # else:
-    flow_obs = flow_obs[warm_up:, :]
-    
-    # Create dataset sample dict.
-    dataset_sample = {
-        'batch_sample': i_sample,
-        'x_phy': select_subset(config, dataset_dict['x_phy'],
-                        i_sample, i_t, config['dpl_model']['rho'],
-                        warm_up=warm_up),
-        'xc_nn_norm': select_subset(
-            config, dataset_dict['xc_nn_norm'], i_sample, i_t,
-            config['dpl_model']['rho'], has_grad=False, warm_up=warm_up),
-        'c_nn': torch.tensor(dataset_dict['c_nn'][i_sample],
-                             device=config['device'], dtype=torch.float32),
-        'target': flow_obs,
-
-    }
-
-    return dataset_sample
-
-
-def get_training_sample_2_0(
-    dataset_dict: Dict[str, np.ndarray], 
-    ngrid_train: int,
-    nt: int,
-    config: Dict,
-) -> Dict[str, torch.Tensor]:
-    """Select random sample of data for training batch."""
-    warm_up = config['dpl_model']['phy_model']['warm_up']
-    nsubbasin_max = config['dpl_model']['phy_model']['nsubbasin_max']
-    subset_dims = (config['train']['batch_size'], config['dpl_model']['rho'])
-
-    i_sample, i_t = random_index(ngrid_train, nt, subset_dims, warm_up=warm_up)
-
-    gage_batch = np.array(dataset_dict['selected_gage'])[i_sample]
-    subbasin_idx = dataset_dict['subbasin_idx']
-
-    id_list = []
-    start_id = 0
-    for gage_idx, gage in enumerate(gage_batch):
-        if(start_id+len(subbasin_idx[gage])) > nsubbasin_max:
-            print("Minibatch size is shrinked to ",gage_idx,".  The maximum number of subbasin is ", start_id+len(subbasin_idx[gage]),f", out of {nsubbasin_max} in the setting ")
-            
-            i_sample=i_sample[:gage_idx]
-            i_t = i_t[:gage_idx]
-            
-            gage_batch = gage_batch[:gage_idx]
-            
-            break
-
-        id_list.append(range(start_id, start_id+len(subbasin_idx[gage])))
-        
-        start_id = start_id+len(subbasin_idx[gage])
-
-        xTrain =  np.full((start_id,config['dpl_model']['rho']+warm_up,dataset_dict['xc_nn_norm'].shape[-1]),np.nan)
-        xTrain2 = np.full((start_id,config['dpl_model']['rho']+warm_up,dataset_dict['xc_nn_norm'].shape[-1]),np.nan)
-        attr2 = np.full((start_id,dataset_dict['c_nn_norm'].shape[-1]),np.nan)
-        
-        idx_matric = np.zeros((start_id,len(gage_batch)))
-        Ai_batch = []
-        Ac_batch = []
-        Ele_batch = []        
-        for gageidx , gage in enumerate(gage_batch):
-
-            idx_matric[np.array(id_list[gageidx]),gageidx] = 1
-            
-            Ai_batch.extend(dataset_dict['Ai_all'][np.array(subbasin_idx[gage]).astype(int)]/np.array(dataset_dict['Ai_all'][np.array(subbasin_idx[gage]).astype(int)]).sum())
-            Ac_batch.extend(dataset_dict['Ac_all'][np.array(subbasin_idx[gage]).astype(int)])
-            Ele_batch.extend(dataset_dict['Ele_all'][np.array(subbasin_idx[gage]).astype(int)])
-            xTrain[np.array(id_list[gageidx]),:,:] = dataset_dict['x_phy'][np.array(subbasin_idx[gage]).astype(int),i_t[gageidx]-warm_up:i_t[gageidx]+config['dpl_model']['rho'],:]
-            xTrain2[np.array(id_list[gageidx]),:,:] = dataset_dict['xc_nn_norm'][np.array(subbasin_idx[gage]).astype(int),i_t[gageidx]-warm_up:i_t[gageidx]+config['dpl_model']['rho'],:]
-            attr2[np.array(id_list[gageidx]),:] = dataset_dict['c_nn_norm'][np.array(subbasin_idx[gage]).astype(int),:]   
-
-        xTrain_torch = torch.from_numpy(np.swapaxes(xTrain, 0,1)).to(config['device'])
-        attr_torch = torch.from_numpy(attr2).to(config['device'])
-
-        attr2_expand = np.repeat(np.expand_dims(attr2, axis=1), xTrain2.shape[1], axis=1)
-        zTrain2 = np.concatenate((xTrain2,attr2_expand),axis = -1)
-        zTrain2_torch = torch.from_numpy(np.swapaxes(zTrain2, 0, 1)).to(config['device'])
-        
-
-
-
-
-    flow_obs = select_subset(config, np.transpose(dataset_dict['target'], (1,0,2)), i_sample, i_t,
-                             config['dpl_model']['rho'], warm_up=warm_up)
-
-    flow_obs = flow_obs[warm_up:, :]
-    
-    # Create dataset sample dict.
-    dataset_sample = {
-        'batch_sample': i_sample,
-        'x_phy': xTrain_torch,
-        'xc_nn_norm': zTrain2_torch,
-        'c_nn_norm': attr_torch,
-        'Ai_batch': Ai_batch,
-        'Ac_batch': Ac_batch,
-        'Ele_batch': Ele_batch,
-        'idx_matric': idx_matric,
-        'target': flow_obs,
-
-    }
-
-    return dataset_sample
-
-
-def _get_validation_sample(
-    dataset_dict: Dict[str, torch.Tensor],
-    i_s: int,
-    i_e: int,
-    config: Dict
-) -> Dict[str, torch.Tensor]:
-    """
-    Take sample of data for testing batch.
-
-    Depreciated.
-    """
-    dataset_sample = {}
-    for key, value in dataset_dict.items():
-        if value.ndim == 3:
-            if key in ['x_phy', 'xc_nn_norm']:
-                warm_up = 0
-            else:
-                warm_up = config['dpl_model']['phy_model']['warm_up']
-            dataset_sample[key] = torch.tensor(
-                value[warm_up:, i_s:i_e, :],
-                dtype=torch.float32,
-                device = config['device']
-            )
-        elif value.ndim == 2:
-            dataset_sample[key] = torch.tensor(
-                value[i_s:i_e, :],
-                dtype=torch.float32,
-                device = config['device']
-            )
-        elif value.ndim == 1:
-            dataset_sample[key] = torch.tensor(
-                value[i_s:i_e],
-                dtype=torch.float32,
-                device = config['device']
-            )
-        else:
-            raise ValueError(f"Incorrect input dimensions. {key} array must have 1, 2 or 3 dimensions.")
-    return dataset_sample
-
-
-def take_sample(config: Dict, dataset_dict: Dict[str, torch.Tensor], days=730,
-                basins=100) -> Dict[str, torch.Tensor]:
-    """Take sample of data."""
-    dataset_sample = {}
-    for key, value in dataset_dict.items():
-        if value.ndim == 3:
-            if key in ['x_phy', 'xc_nn_norm']:
-                warm_up = 0
-            else:
-                warm_up = config['dpl_model']['phy_model']['warm_up']
-            dataset_sample[key] = torch.tensor(value[warm_up:days,:basins, :], dtype=torch.float32).to(config['device'])
-        elif value.ndim == 2:
-            dataset_sample[key] = torch.tensor(value[:basins, :], dtype=torch.float32).to(config['device'])
-        else:
-            raise ValueError(f"Incorrect input dimensions. {key} array must have 2 or 3 dimensions.")
-
-    # Keep 'warmup' days for dHBV1.1p.
-    if ('HBV1_1p' in config['dpl_model']['phy_model']['model']) and \
-    (config['dpl_model']['phy_model']['use_warmup_mode']) and (config['multimodel_type'] == 'none'):
-        pass
-    else:
-        dataset_sample['target'] = dataset_sample['target'][config['dpl_model']['phy_model']['warm_up']:days, :basins]
-    return dataset_sample
-
-
-def numpy_to_torch_dict(data_dict: Dict[str, 
-np.ndarray], device: str) -> Dict[str, torch.Tensor]:
+def numpy_to_torch_dict(
+        data_dict: Dict[str, np.ndarray],
+        device: str
+    ) -> Dict[str, torch.Tensor]:
     """Convert numpy data dictionary to torch tensor dictionary.
 
     Parameters
