+++ deltaModel/core/data/data_samplers/sampler_hydro_ms.py
@@ -3,7 +3,6 @@ from typing import Dict, Optional
 import numpy as np
 import torch
 
-from core.data import random_index
 from core.data.data_samplers.base import BaseDataSampler
