+++ deltaModel/models/loss_functions/kge_norm_loss_batch.py
@@ -1,8 +1,6 @@
 from typing import Any, Dict, Optional
 
-import numpy as np
 import torch
-from numpy.typing import NDArray
 
 
 class KgeNormLossBatch(torch.nn.Module):
