+++ deltaModel/models/loss_functions/nse_loss_batch.py
@@ -2,7 +2,6 @@ from typing import Any, Dict, Optional
 
 import numpy as np
 import torch
-from numpy.typing import NDArray
 
 
 class NseLossBatch(torch.nn.Module):
