+++ deltaModel/models/loss_functions/rmse_loss_comb.py
@@ -1,8 +1,6 @@
 from typing import Any, Dict, Optional
 
-import numpy as np
 import torch
-from numpy.typing import NDArray
 
 
 class RmseLossComb(torch.nn.Module):
