+++ deltaModel/models/multimodel/ensemble_generator.py
@@ -1,4 +1,3 @@
-from math import log
 from typing import Any, Dict, List, Optional, Tuple
 
 import torch
