+++ deltaModel/trainers/trainer_ms.py
@@ -193,7 +193,7 @@ class Trainer(BaseTrainer):
 
         # Track overall predictions and observations
         batch_predictions = []
-        if  self.config['test']['evaluation']: 
+        if self.config['test']['evaluation']: 
             observations = self.eval_dataset['target']
         else:
             observations = None
@@ -239,7 +239,7 @@ class Trainer(BaseTrainer):
 
         # Get start and end indices for each batch.
         n_samples = self.inf_dataset['xc_nn_norm'].shape[1]
-        batch_start = np.arange(0, n_samples, self.config['test']['batch_size'])
+        batch_start = np.arange(0, n_samples, self.config['predict']['batch_size'])
         batch_end = np.append(batch_start[1:], n_samples)
 
         # Forward loop
