--- example/hydrology/trained_models/statistics_basinnorm.json
@@ -1,236 +1,0 @@
-{
-    "flow_sim": [
-        -0.6539479346221658,
-        0.3041285235687269,
-        -0.1374393360006698,
-        0.3625230576307994
-    ],
-    "prcp": [
-        0.0,
-        11.18,
-        3.296422257962528,
-        7.891161606628799
-    ],
-    "tmean": [
-        -3.0,
-        24.049999999999997,
-        10.673338012793195,
-        10.244311069959249
-    ],
-    "pet": [
-        0.59073005,
-        5.52216017,
-        2.922584956632652,
-        1.8546816701101507
-    ],
-    "p_mean": [
-        1.77527036276523,
-        5.0380848733744,
-        3.3940368166168464,
-        1.3664790514127552
-    ],
-    "pet_mean": [
-        2.15246869267625,
-        3.53751126625599,
-        2.7699391292357896,
-        0.5440039229998542
-    ],
-    "p_seasonality": [
-        -0.816394543593238,
-        0.570789214497895,
-        -0.03238007058832508,
-        0.49217497288717404
-    ],
-    "frac_snow": [
-        0.0037075564040486,
-        0.513670726431351,
-        0.1739066705321371,
-        0.20131414745525925
-    ],
-    "aridity": [
-        0.523147061580375,
-        1.68094218186979,
-        0.9738483237395337,
-        0.512143027719942
-    ],
-    "high_prec_freq": [
-        13.9,
-        25.7,
-        20.80225988700565,
-        4.414910596023075
-    ],
-    "high_prec_dur": [
-        1.16371681415929,
-        1.60119047619048,
-        1.3366068671990068,
-        0.18940408064956685
-    ],
-    "low_prec_freq": [
-        204.05,
-        297.05,
-        251.83474576271186,
-        33.61013717586647
-    ],
-    "low_prec_dur": [
-        3.37646198830409,
-        8.6919770773639,
-        5.6039276133210185,
-        2.542981743541223
-    ],
-    "elev_mean": [
-        109.86,
-        2113.28,
-        728.0364218455743,
-        787.0313661469036
-    ],
-    "slope_mean": [
-        4.53352,
-        116.4248,
-        46.185522467043306,
-        47.21662736978322
-    ],
-    "area_gages2": [
-        41.72,
-        1179.59,
-        477.2376271186441,
-        472.4784958241116
-    ],
-    "frac_forest": [
-        0.0386,
-        0.9972,
-        0.6665877589453859,
-        0.36137890520476257
-    ],
-    "lai_max": [
-        1.33837641261247,
-        5.06947386160055,
-        3.338727284241188,
-        1.4742730161996775
-    ],
-    "lai_diff": [
-        0.838265114689036,
-        4.3210609777145,
-        2.5575539919469894,
-        1.3155213847139586
-    ],
-    "gvf_max": [
-        0.504651094366985,
-        0.887158068880153,
-        0.7406071166056095,
-        0.15453392576777641
-    ],
-    "gvf_diff": [
-        0.135148907505314,
-        0.508506796279639,
-        0.3333465879997043,
-        0.1482936757653723
-    ],
-    "dom_land_cover_frac": [
-        0.539183448712223,
-        0.999999999999898,
-        0.8101649488307202,
-        0.18493105874108887
-    ],
-    "dom_land_cover": [
-        3.0,
-        11.0,
-        6.135593220338983,
-        3.013540468399415
-    ],
-    "root_depth_50": [
-        0.1363996871346371,
-        0.2254880993297035,
-        0.18070237751968754,
-        0.030519557407959627
-    ],
-    "soil_depth_pelletier": [
-        0.800724637681159,
-        41.9182968929804,
-        10.307206608611155,
-        16.03859634489692
-    ],
-    "soil_depth_statsgo": [
-        0.884824261732569,
-        1.50000001490116,
-        1.295058160463927,
-        0.2579013007557458
-    ],
-    "soil_porosity": [
-        0.416396539872731,
-        0.46773459992883,
-        0.4425870926995773,
-        0.021234984278231863
-    ],
-    "soil_conductivity": [
-        0.650205789053475,
-        2.84356441326312,
-        1.6578373700435987,
-        1.3300314726169684
-    ],
-    "max_water_content": [
-        0.348778891917816,
-        0.683067168114119,
-        0.5308235232377105,
-        0.13901835583212488
-    ],
-    "sand_frac": [
-        16.8490255147272,
-        57.3021340831677,
-        36.09920885000754,
-        15.729734143027096
-    ],
-    "silt_frac": [
-        16.9619709076376,
-        51.8799998656123,
-        34.77536146479991,
-        13.247512672179738
-    ],
-    "clay_frac": [
-        8.68936820600364,
-        32.9897292954575,
-        20.364762937719206,
-        9.298753460157364
-    ],
-    "geol_1st_class": [
-        3.0,
-        11.0,
-        7.60075329566855,
-        3.274440105007857
-    ],
-    "glim_1st_class_frac": [
-        0.503876589054494,
-        1.0,
-        0.7932291830054035,
-        0.19856030925283152
-    ],
-    "geol_2nd_class": [
-        -1.0,
-        11.0,
-        5.284369114877589,
-        4.6674858570044035
-    ],
-    "glim_2nd_class_frac": [
-        0.0,
-        0.356563630868354,
-        0.14909449785494944,
-        0.1395197189996488
-    ],
-    "carbonate_rocks_frac": [
-        0.0,
-        0.636085186630631,
-        0.12765520210330902,
-        0.27170019368155496
-    ],
-    "geol_porosity": [
-        0.017840000000000005,
-        0.21416000000000002,
-        0.12352518939393939,
-        0.0686217237858604
-    ],
-    "geol_permeability": [
-        -15.466,
-        -12.4039,
-        -13.94927570621469,
-        1.1820558801502403
-    ]
-}
